fix ExportOrder + queues

See merge request hai95.lv/Unicoin-Clothing-Website!63

+++ product-service/src/main/java/com/unicoin/product/dto/ExportOrderDetailDTO.java
@@ -15,7 +15,7 @@ public class ExportOrderDetailDTO {
 
     private Long id;
 
-    private Integer variantId;
+    private Long variantId;
 
     private String variantName;
